# docs/swagger.yaml
basePath: /
definitions:
...
  requests.LoanCheckAvailabilityPayload:
    properties:
      book_id:
        type: integer
      student_id:
        type: integer
    type: object
  requests.LoanPayload:
    properties:
      book_id:
        type: integer
      due_date:
        type: string
      loan_date:
        type: string
      notes:
        type: string
      student_id:
        type: integer
    type: object
  requests.LoanUpdatePayload:
    properties:
      due_date:
        type: string
      notes:
        type: string
      status:
        type: string
    type: object
...
  resources.AuthorResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.BookResource:
    properties:
      author:
        $ref: '#/definitions/resources.AuthorResource'
      author_id:
        type: integer
      barcode:
        type: string
      created_at:
        type: string
      fixture_no:
        type: integer
      id:
        type: integer
      is_currently_loaned:
        type: boolean
      is_donation:
        type: boolean
      language:
        type: string
      level:
        type: string
      name:
        type: string
      page_count:
        type: integer
      publisher:
        $ref: '#/definitions/resources.PublisherResource'
      publisher_id:
        type: integer
      shelf_code:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.ClassroomResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  resources.LoanCollection:
    properties:
      data:
        items:
          $ref: '#/definitions/resources.LoanResource'
        type: array
      meta:
        $ref: '#/definitions/resources.PaginationMeta'
    type: object
  resources.LoanResource:
    properties:
      book:
        $ref: '#/definitions/resources.BookResource'
      book_id:
        type: integer
      created_at:
        type: string
      due_date:
        type: string
      id:
        type: integer
      loan_date:
        type: string
      notes:
        type: string
      return_date:
        type: string
      status:
        type: string
      student:
        $ref: '#/definitions/resources.StudentResource'
      student_id:
        type: integer
      updated_at:
        type: string
    type: object
  resources.PaginationMeta:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
...
  resources.PublisherResource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  resources.StudentResource:
    properties:
      classroom:
        $ref: '#/definitions/resources.ClassroomResource'
      classroom_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.ErrorBody:
    properties:
      details: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  title: Safa API
  version: "1.0"
paths:
...
  /loans:
    get:
      description: Get a paginated list of loans with optional filters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 15)
        in: query
        name: per_page
        type: integer
      - description: Filter by status (active/returned/overdue)
        in: query
        name: status
        type: string
      - description: Search by student name
        in: query
        name: search
        type: string
      - description: Filter by student ID
        in: query
        name: student_id
        type: integer
      - description: Filter by book ID
        in: query
        name: book_id
        type: integer
      - description: Sort column
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.LoanCollection'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: List all Loan records
      tags:
      - Loan
    post:
      consumes:
      - application/json
      description: Creates a new loan after checking level compatibility and active-loan
        constraints.
      parameters:
      - description: Loan payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.LoanPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resources.LoanResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a Loan
      tags:
      - Loan
  /loans/{id}:
    delete:
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete a Loan
      tags:
      - Loan
    get:
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.LoanResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get a Loan by ID
      tags:
      - Loan
    put:
      consumes:
      - application/json
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Loan update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.LoanUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.LoanResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update a Loan
      tags:
      - Loan
  /loans/{id}/return:
    post:
      parameters:
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.LoanResource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorBody'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Mark a loan as returned
      tags:
      - Loan
  /loans/check-availability:
    post:
      consumes:
      - application/json
      parameters:
      - description: Check payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.LoanCheckAvailabilityPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorBody'
      security:
      - BearerAuth: []
      summary: Check if a student can borrow a book
      tags:
      - Loan
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"